+++ CHANGELOG.md
@@ -48,6 +48,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `burning` condition
 - `inconversation` condition
 - `heroesattribute` condition - Heroes compatibility feature: Checks a player's level for a particular attribute against a value
+- `mmspawn` event now has argument `target` Makes the spawned mob target the player
+- `mmspawn` event now has argument `private` Visually hides the spawned mob from other players. Does not stop sound or particles
+- `mmspawn` event now supports the `marked` argument
+- `mmobkill` objective now supports the `marked` argument
 - Things that are also added in 1.12.X:
     - new line support for `journal_lore` in `messages.yml`
     - FastAsyncWorldEdit compatibility